Implement function!() macro, v3
v3: names from static record initializers should not be appended,
and module names should also not be prepended because they are
already covered by module_path macro.

v2: fix for trait default methods, impl methods, and take the X
from surrounding 'impl X {}' for 'impl _ for X {}'.

In addition to a line!() and module_path!(), I really miss the ability
from C to receive the name of the current function as a string. There's
some discussion on how to accomplish this and what to return, and I think
it should by minimum include the function names, without the modules.

For example,

test.rs:

    fn main()
    {
        fn inner() {
            println!("{}", function!());
            fn inner_a() {
                println!("{}", function!());
            }
            inner_a();
        }

        fn inner2() {
            println!("{}", function!());
            fn inner_b() {
                println!("{}", function!());
            }
            inner_b();
        }

        println!("{}", module_path!());
        println!("{}", function!());
        inner();
        inner2();
    ]

Emits:

    test
    main
    main::inner
    main::inner::inner_a
    main::inner2
    main::inner2::inner_b

Signed-off-by: Dan Aloni <alonid@gmail.com>